avoiding movie not found error

+++ backend/src/service/MovieService/DeleteMovieService/index.ts
@@ -6,12 +6,21 @@ const prisma = new PrismaClient();
 export class DeleteMovieService {
   async execute( req: Request, res: Response) {
     const { id } = req.params;
-
-    await prisma.movie.delete({
+    
+    const movie = await prisma.movie.findUnique({
       where: {
         id: Number(id),
       },
     });
+    if (movie){
+      await prisma.movie.delete({
+        where: {
+          id: Number(id),
+        },
+      });
+    } else {
+      return res.json({ Message: "Filme inexistente" });
+    }
 
     return res.json({ Message: "Filme removido com sucesso!" });
   }
